Consolidate analytics reporting which was duplicated on 3 places
