+++ plugins/src/main/java/com/jraska/analytics/AnalyticsEvent.kt
@@ -1,0 +1,6 @@
+package com.jraska.analytics
+
+data class AnalyticsEvent(
+  val name: String,
+  val properties: Map<String, Any?>
+)
