+++ plugins/src/main/java/com/jraska/analytics/AnalyticsReporter.kt
@@ -1,0 +1,19 @@
+package com.jraska.analytics
+
+import com.mixpanel.mixpanelapi.MixpanelAPI
+
+interface AnalyticsReporter {
+  fun report(vararg events: AnalyticsEvent)
+
+  companion object {
+    fun create(reporterName: String): AnalyticsReporter {
+      val mixpanelToken: String? = System.getenv("GITHUB_CLIENT_MIXPANEL_API_KEY")
+      if (mixpanelToken == null) {
+        println("'GITHUB_CLIENT_MIXPANEL_API_KEY' not set, data will be reported to console only")
+        return ConsoleReporter(reporterName)
+      } else {
+        return MixpanelAnalyticsReporter(mixpanelToken, MixpanelAPI(), reporterName)
+      }
+    }
+  }
+}
