+++ plugins/src/main/java/com/jraska/analytics/ConsoleReporter.kt
@@ -1,0 +1,9 @@
+package com.jraska.analytics
+
+class ConsoleReporter(private val reporterName: String) : AnalyticsReporter {
+  override fun report(vararg events: AnalyticsEvent) {
+    events.forEach {
+      println("$reporterName: $it")
+    }
+  }
+}
