--- plugins/src/main/java/com/jraska/github/client/firebase/TestResultReporter.kt
@@ -1,5 +1,0 @@
-package com.jraska.github.client.firebase
-
-interface TestResultReporter {
-  fun report(results: TestSuiteResult)
-}
