--- plugins/src/main/java/com/jraska/github/client/firebase/report/ConsoleTestResultReporter.kt
@@ -1,10 +1,0 @@
-package com.jraska.github.client.firebase.report
-
-import com.jraska.github.client.firebase.TestResultReporter
-import com.jraska.github.client.firebase.TestSuiteResult
-
-class ConsoleTestResultReporter : TestResultReporter {
-  override fun report(results: TestSuiteResult) {
-    println(results)
-  }
-}
