--- plugins/src/main/java/com/jraska/gradle/buildtime/report/ConsoleReporter.kt
@@ -1,10 +1,0 @@
-package com.jraska.gradle.buildtime.report
-
-import com.jraska.gradle.buildtime.BuildData
-import com.jraska.gradle.buildtime.BuildReporter
-
-class ConsoleReporter : BuildReporter {
-  override fun report(buildData: BuildData) {
-    println(buildData)
-  }
-}
